requirements.txt
fitz
PyMuPDF
gradio
numpy
openai
pymongo
python-dotenv
sentence_transformers
